Restructure project packages according to routing
