+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projects/project/experiments/ExperimentalDesignCard.java
@@ -1,10 +1,11 @@
-package life.qbic.datamanager.views.project.view.pages.experiment.components;
+package life.qbic.datamanager.views.projects.project.experiments;
 
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import life.qbic.datamanager.views.layouts.CardLayout;
+import life.qbic.datamanager.views.projects.project.ProjectViewPage;
 import life.qbic.projectmanagement.domain.project.experiment.Experiment;
 
 /**
@@ -12,8 +13,7 @@ import life.qbic.projectmanagement.domain.project.experiment.Experiment;
  *
  * <p>A CardLayout based Component showing the information stored in the
  * {@link life.qbic.projectmanagement.domain.project.experiment.ExperimentalDesign} associated with
- * a project in the {@link ExperimentListComponent} of the
- * {@link life.qbic.datamanager.views.project.view.ProjectViewPage}
+ * a project in the {@link ExperimentListComponent} of the {@link ProjectViewPage}
  */
 public class ExperimentalDesignCard extends CardLayout {
